requirements.txt
kafka-python==2.0.2
faker==25.0.0
pyspark==3.5.1
psycopg2-binary==2.9.9
